Move transformArraysIntoObjects helper into a separate module
